Move Sample to correct module
