+++ ccd-sample-data/src/main/java/uk/gov/hmcts/cmc/ccd/sample/data/SampleCCDTelephone.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.cmc.ccd.util;
+package uk.gov.hmcts.cmc.ccd.sample.data;
 
 import uk.gov.hmcts.cmc.ccd.domain.CCDTelephone;
